+++ src/com/intervigil/micdroid/AutoTalent.java
@@ -11,7 +11,7 @@ public class AutoTalent {
 	
 	public static native void initializeAutoTalent(float concertA, char key, 
 			float fixedPitch, float fixedPull, 
-			float correctStrength, float correctSmooth, float pitchShift, float scaleRotate,
+			float correctStrength, float correctSmooth, float pitchShift, int scaleRotate,
 			float lfoDepth, float lfoRate, float lfoShape, float lfoSym, int lfoQuant, 
 			int formCorr, float formWarp, float mix);
